Move the `social` array to where it is used.
